Create fromanAi-pleasehelp
This is just an attempt to build a objectenetic agent with 10 objects they define themselves with from an 32 x 32 array with a 0-9 value at each point. Eventually the environment will have a mechanism to allow agents to couple and make new agents. While half they info will pass on and remain the same in the new agent. The rules for coupling will be simple and eventually the spots that the total of them minus the 100 will be what defines all the possibilities.