Create start_of_an_environment
Well what I want to do is expound upon this or other environments. I have some idea form an environment there is a cot to coupling but it is less costly then not coupling except if the agents generate an undesirable agent to any other agent or has rows or columns that are the same, meaning 2 spots in one or the other. Disability comes from the 10 spots possibly matching the with 5 random one coming form each of the parents. Then every time an agents couples the ancestor couples gets points in a "reward" bank. this is filled with an initial amount like 100 and then the females have 20 turns they can couple and males 60 turns, when you don't couple but can you loose 2 from the reward bank, when you cannot couple any more you still may receive reward points if you have descendants coupling. When you couple your offspring costs you 1 each parent each turn in addition to the other costs, but when it couples this cost is passed on to it and the ancestors are rewarded with a +3. When the turns are no more that you can couple the cost of each turn is -4 per agent. The turns are linear and the rewards bank is like the life. When it is empty the agent dies.