Create another_enviroment_maybe
Eventually I want to have agents that have 10 objects each and exchange 5 spots form each of their objects with another agent. these 5 spot are random from the 10 spot available and there is 10 objects per agent. The desire to couple needs to be something it has a cost to couple but if a desirable offspring is the product of coupling then the rewards can be filled. Getting the exchange of the objectenetical info is important. Then having basic rules for different behaviors can be we established and then a attributes of the agents can be made from all of the 1024 x 10 spots - the 100 used to identify agents. So an agent will be has veritable as there are "spots" unused and there locations. the location and the compatibility is rows and columns with shared spots carry an -10% per deficit of row and column. When offspring don't couple they cost and this taxes the agents that coupled and produces undesirable agents. Disability is a measure of how many of the 0 are filled with 9 or closer to than a 0 or spot marked with a x or y on the X column.